camerad: cleanup DM exposure and move into camera_qcom2.cc (#24391)
* camerad: cleanup DM exposure and move into camera_qcom2.cc

* remove include
